+++ envision/tests/test_data_replay.py
@@ -33,7 +33,7 @@ from smarts.core.agent_interface import AgentInterface, AgentType
 from smarts.core.scenario import Scenario
 from smarts.core.smarts import SMARTS
 from smarts.core.sumo_traffic_simulation import SumoTrafficSimulation
-from envision.utils.queue import Queue
+from envision.utils.multiprocessing_queue import Queue
 
 AGENT_ID = "Agent-007"
 NUM_EPISODES = 3
